pyproject.toml
[tool.poetry]
name = "l3-fastapi-osmthome-backend"
version = "0.1.0"
description = "OSMTHOME BACKEND"
authors = ["KamruzzamanAsif <asif720kamruzzaman@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
pre-commit = "^3.6.2"
pyodbc = "^5.1.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
app = "src.main:start"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
